Add more integration tests for create category page and createproduct page
