+++ .env
@@ -5,3 +5,8 @@ JWT_SECRET = HGFHGEAD12124322432
 BRAINTREE_MERCHANT_ID = hmrc3kfrt2xrvtvp
 BRAINTREE_PUBLIC_KEY = d3rnqcjwn7zk4fpt
 BRAINTREE_PRIVATE_KEY = 767ed9ddd903781349d4c2af441f8eaa
+TEST_ADMIN_USER = admin@test.sg
+TEST_ADMIN_PASS = admin@test.sg
+TEST_NORMAL_USER = cs4218@test.com
+TEST_NORMAL_PASS = cs4218@test.com
+
