+++ client/src/components/Routes/AdminRoute.js
@@ -1,25 +1,26 @@
-import { useState,useEffect } from "react";
+import React from "react";
+import { useState, useEffect } from "react";
 import { useAuth } from "../../context/auth";
 import { Outlet } from "react-router-dom";
-import axios from 'axios';
+import axios from "axios";
 import { set } from "mongoose";
 import Spinner from "../Spinner";
 
-export default function AdminRoute(){
-    const [ok,setOk] = useState(false)
-    const [auth,setAuth] = useAuth()
+export default function AdminRoute() {
+  const [ok, setOk] = useState(false);
+  const [auth, setAuth] = useAuth();
 
-    useEffect(()=> {
-        const authCheck = async() => {
-            const res = await axios.get("/api/v1/auth/admin-auth");
-            if(res.data.ok){
-                setOk(true);
-            } else {
-                setOk(false);
-            }
-        };
-        if (auth?.token) authCheck();
-    }, [auth?.token]);
-    
-    return ok ? <Outlet /> : <Spinner/>;
+  useEffect(() => {
+    const authCheck = async () => {
+      const res = await axios.get("/api/v1/auth/admin-auth");
+      if (res.data.ok) {
+        setOk(true);
+      } else {
+        setOk(false);
+      }
+    };
+    if (auth?.token) authCheck();
+  }, [auth?.token]);
+
+  return ok ? <Outlet /> : <Spinner />;
 }
